+++ models/web/requests/wallet_update_request.go
@@ -5,6 +5,6 @@ type WalletUpdateRequest struct {
 	Name     string `validate:"required,max=200,min=1" json:"name"`
 	Icon     string `json:"icon"`
 	Currency string `validate:"required" json:"currency"`
-	Balance  int32  `json:"balance"`
+	Balance  int64  `json:"balance"`
 	UserID   int32  `validate:"required" json:"user_id"`
 }
