+++ models/web/responses/wallet_response.go
@@ -7,7 +7,7 @@ type WalletResponse struct {
 	Name      string    `json:"name"`
 	Icon      string    `json:"icon"`
 	Currency  string    `json:"currency"`
-	Balance   int32     `json:"balance"`
+	Balance   int64     `json:"balance"`
 	UserID    int32     `json:"user_id"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
